Move agent and workflow transport internals into services
Expose agent provider transport primitives from services/agents and workflow provider transport primitives from services/workflows while keeping internal/providerhost as a compatibility wrapper.\n\nShared wire helpers now live under services/internal so workflows can encode workflow agent targets without copying agent message/tool-ref conversion logic.\n\nNew interfaces/examples:\n\n    agentProvider, err := agents.NewExecutable(ctx, agents.ExecConfig{\n        Command: command,\n        Args: args,\n        Config: config,\n        HostServices: hostServices,\n        Name: providerName,\n        Telemetry: telemetry,\n    })\n\n    agentProvider, err := agents.NewRemote(ctx, agents.RemoteConfig{\n        Client: proto.NewAgentProviderClient(conn),\n        Runtime: proto.NewProviderLifecycleClient(conn),\n        Name: providerName,\n    })\n\n    proto.RegisterAgentHostServer(srv, agents.NewHostServer(providerName, searchTools, executeTool))\n    proto.RegisterAgentManagerHostServer(srv, agents.NewManagerServer(pluginName, manager, tokens))\n\n    workflowProvider, err := workflows.NewExecutable(ctx, workflows.ExecConfig{\n        Command: command,\n        Args: args,\n        Config: config,\n        HostServices: hostServices,\n        Name: providerName,\n    })\n\n    proto.RegisterWorkflowHostServer(srv, workflows.NewHostServer(providerName, invoke))\n    proto.RegisterWorkflowManagerHostServer(srv, workflows.NewManagerServer(pluginName, manager, tokens))\n\nCompatibility examples remain available:\n\n    providerhost.NewExecutableAgent(ctx, providerhost.AgentExecConfig{...})\n    providerhost.NewExecutableWorkflow(ctx, providerhost.WorkflowExecConfig{...})\n\nThe moved contract/transport tests now live with the service packages so providerhost no longer owns these internals.
