+++ gestaltd/services/agents/grpc_test.go
@@ -1,4 +1,4 @@
-package providerhost
+package agents
 
 import (
 	"context"
@@ -10,11 +10,11 @@ import (
 	"google.golang.org/grpc/test/bufconn"
 )
 
-func newBufconnConnWithOptions(t *testing.T, serverOpts []grpc.ServerOption, dialOpts []grpc.DialOption, register func(*grpc.Server)) *grpc.ClientConn {
+func newBufconnConn(t *testing.T, register func(*grpc.Server)) *grpc.ClientConn {
 	t.Helper()
 
 	lis := bufconn.Listen(1024 * 1024)
-	srv := grpc.NewServer(serverOpts...)
+	srv := grpc.NewServer()
 	register(srv)
 	go func() {
 		_ = srv.Serve(lis)
@@ -24,14 +24,12 @@ func newBufconnConnWithOptions(t *testing.T, serverOpts []grpc.ServerOption, dia
 		_ = lis.Close()
 	})
 
-	dialOptions := []grpc.DialOption{
+	conn, err := grpc.NewClient("passthrough:///bufnet",
 		grpc.WithContextDialer(func(context.Context, string) (net.Conn, error) {
 			return lis.Dial()
 		}),
 		grpc.WithTransportCredentials(insecure.NewCredentials()),
-	}
-	dialOptions = append(dialOptions, dialOpts...)
-	conn, err := grpc.NewClient("passthrough:///bufnet", dialOptions...)
+	)
 	if err != nil {
 		t.Fatalf("grpc.NewClient: %v", err)
 	}
