+++ gestaltd/services/agents/provider_lifecycle_fake_test.go
@@ -1,4 +1,4 @@
-package providerhost
+package agents
 
 import (
 	"context"
