+++ gestaltd/services/workflows/client_test.go
@@ -1,4 +1,4 @@
-package providerhost
+package workflows
 
 import (
 	"context"
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/valon-technologies/gestalt/server/internal/egress"
+	"github.com/valon-technologies/gestalt/server/services/runtimehost"
 )
 
 func TestNewExecutableWorkflowForwardsDefaultEgressAction(t *testing.T) {
@@ -15,18 +16,18 @@ func TestNewExecutableWorkflowForwardsDefaultEgressAction(t *testing.T) {
 	defer func() { startWorkflowProviderProcess = originalStart }()
 
 	wantErr := errors.New("boom")
-	var got ProcessConfig
-	startWorkflowProviderProcess = func(_ context.Context, cfg ProcessConfig) (*providerProcess, error) {
+	var got runtimehost.ProcessConfig
+	startWorkflowProviderProcess = func(_ context.Context, cfg runtimehost.ProcessConfig) (*runtimehost.PluginProcess, error) {
 		got = cfg
 		return nil, wantErr
 	}
 
-	_, err := NewExecutableWorkflow(context.Background(), WorkflowExecConfig{
+	_, err := NewExecutable(context.Background(), ExecConfig{
 		Command: "/bin/true",
 		Egress:  egress.Policy{DefaultAction: egress.PolicyDeny},
 	})
 	if !errors.Is(err, wantErr) {
-		t.Fatalf("NewExecutableWorkflow error = %v, want %v", err, wantErr)
+		t.Fatalf("NewExecutable error = %v, want %v", err, wantErr)
 	}
 	if got.Egress.DefaultAction != egress.PolicyDeny {
 		t.Fatalf("default action = %q, want %q", got.Egress.DefaultAction, egress.PolicyDeny)
