+++ gestaltd/services/workflows/convert_workflow_test.go
@@ -1,4 +1,4 @@
-package providerhost
+package workflows
 
 import (
 	"testing"
